+++ codec/console_reader.go
@@ -38,6 +38,8 @@ import (
 	"github.com/streamingfast/logging"
 	"go.uber.org/zap"
 	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/types/known/anypb"
+	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 // ConsoleReader is what reads the `geth` output directly. It builds
@@ -164,8 +166,9 @@ func (s *parsingStats) inc(key string) {
 }
 
 type parseCtx struct {
-	blockVersion int32
-	fhVersion    string
+	blockVersion  int32
+	fhVersion     string
+	useReadBlock2 bool
 
 	currentBlock         *pbeth.Block
 	currentTrace         *pbeth.TransactionTrace
@@ -245,6 +248,9 @@ func (c *ConsoleReader) next(readType int) (out interface{}, err error) {
 		switch {
 		case strings.HasPrefix(line, "BLOCK"):
 			ctx.stats.inc("BLOCK")
+			if ctx.useReadBlock2 {
+				return ctx.readBlock2(line)
+			}
 			return ctx.readBlock(line)
 
 		case strings.HasPrefix(line, "GAS_CHANGE"):
@@ -973,19 +979,32 @@ func (ctx *parseCtx) readCreateAccount(line string) error {
 }
 
 /*
-FIRE INIT 2.0 polygon 1.10.17-fh+hotfix (deadbeef) ...
+current geth instrumented nodes: FIRE INIT 2.x polygon 1.10.17-fh+hotfix (deadbeef) ...
+new rpc poller:                  FIRE INIT 1.0 sf.ethereum.v1.block
+future poller/trace-geth:        FIRE INIT 3.0 sf.ethereum.v1.block 1.10.17-fh+hotfix (deadbeef) ...
 */
 func (ctx *parseCtx) readInit(line string) error {
-	chunks, err := SplitInBoundedChunks(line, 4)
-	if err != nil {
-		return fmt.Errorf("split: %s", err)
-	}
 
+	chunks := strings.SplitN(line, " ", -1)
+	chunks = chunks[1:]
+
+	var nodeVariant, nodeVersion string
+	switch len(chunks) {
+	case 0, 1:
+		return fmt.Errorf("not enough terms in INIT Line: %q", line)
+	case 2:
+		nodeVersion = chunks[1]
+		nodeVariant = "unknown"
+	default:
+		nodeVariant = strings.ToLower(chunks[1])
+		nodeVersion = strings.Join(chunks[2:], " ")
+	}
 	ctx.fhVersion = chunks[0]
-	nodeVariant := strings.ToLower(chunks[1])
-	nodeVersion := chunks[2]
 
 	switch ctx.fhVersion {
+	case "1.0": // 1.0 is erroneously used by the first implementation of the rpc poller. will upgrade to 3.0 in next firecore releases
+		ctx.blockVersion = 3
+		ctx.useReadBlock2 = true
 	case "2.0":
 		ctx.blockVersion = 2
 		ctx.normalizationFeatures.UpgradeBlockV2ToV3 = true
@@ -994,6 +1013,9 @@ func (ctx *parseCtx) readInit(line string) error {
 	case "2.3":
 		ctx.blockVersion = 3
 		ctx.normalizationFeatures.ReorderTransactionsAndRenumberOrdinals = true
+	case "3.0":
+		ctx.blockVersion = 3
+		ctx.useReadBlock2 = true
 	default:
 		return fmt.Errorf("major version of Firehose exchange protocol is unsupported (expected: one of [2.0, 2.1, 2.2, 2.3], found %s), you are most probably running an incompatible version of the Firehose instrumented 'geth' client", ctx.fhVersion)
 	}
@@ -1087,6 +1109,60 @@ func (ctx *parseCtx) readCodeChange(line string) error {
 	evmCall.CodeChanges = append(evmCall.CodeChanges, codeChange)
 
 	return nil
+}
+
+// readBlock2 reads the new format of blocks, the one exported by rpc pollers and tracer-based instrumented geth
+// [block_num:342342342] [block_hash] [parent_num] [parent_hash] [lib:123123123] [timestamp:unix_nano] B64ENCODED_any
+func (ctx *parseCtx) readBlock2(line string) (*pbbstream.Block, error) {
+	chunks, err := SplitInBoundedChunks(line, 8)
+	if err != nil {
+		return nil, fmt.Errorf("splitting block log line: %w", err)
+	}
+
+	blockNum, err := strconv.ParseUint(chunks[0], 10, 64)
+	if err != nil {
+		return nil, fmt.Errorf("parsing block num %q: %w", chunks[0], err)
+	}
+
+	blockHash := chunks[1]
+
+	parentNum, err := strconv.ParseUint(chunks[2], 10, 64)
+	if err != nil {
+		return nil, fmt.Errorf("parsing parent num %q: %w", chunks[2], err)
+	}
+
+	parentHash := chunks[3]
+
+	libNum, err := strconv.ParseUint(chunks[4], 10, 64)
+	if err != nil {
+		return nil, fmt.Errorf("parsing lib num %q: %w", chunks[4], err)
+	}
+
+	timestampUnixNano, err := strconv.ParseUint(chunks[5], 10, 64)
+	if err != nil {
+		return nil, fmt.Errorf("parsing timestamp %q: %w", chunks[5], err)
+	}
+
+	timestamp := time.Unix(0, int64(timestampUnixNano))
+
+	payload, err := base64.StdEncoding.DecodeString(chunks[6])
+
+	blockPayload := &anypb.Any{
+		TypeUrl: "type.googleapis.com/sf.ethereum.type.v2.Block",
+		Value:   payload,
+	}
+
+	block := &pbbstream.Block{
+		Id:        blockHash,
+		Number:    blockNum,
+		ParentId:  parentHash,
+		ParentNum: parentNum,
+		Timestamp: timestamppb.New(timestamp),
+		LibNum:    libNum,
+		Payload:   blockPayload,
+	}
+
+	return block, nil
 }
 
 // Formats